requirements.txt
PyQt5>=5.15.0
pyserial>=3.4
RPi.GPIO>=0.7.0
